Feature-#12 파트 레이아웃 구현 - 파트 레이아웃 생성과 화면전환
