+++ lib/nav_body.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
-import 'package:word_english/screen/community/s_community.dart';
-import 'package:word_english/screen/home/s_home.dart';
-import 'package:word_english/screen/my/s_my.dart';
-import 'package:word_english/screen/others/s_others.dart';
+import 'package:word_english/screen/tabs/community/s_community.dart';
+import 'package:word_english/screen/tabs/home/s_home.dart';
+import 'package:word_english/screen/tabs/my/s_my.dart';
+import 'package:word_english/screen/tabs/others/s_others.dart';
 
 class NavBody extends StatelessWidget {
   final int index;
