+++ lib/screen/tabs/home/fragments/f_part.dart
@@ -1,0 +1,14 @@
+import 'package:flutter/material.dart';
+import 'package:provider/provider.dart';
+import 'package:velocity_x/velocity_x.dart';
+
+import '../../../../provider/package_info_provider.dart';
+
+class PartFragment extends StatelessWidget {
+  const PartFragment({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return Placeholder();
+  }
+}
