+++ lib/word_english_app.dart
@@ -1,18 +1,26 @@
 import 'package:flutter/material.dart';
+import 'package:nav/nav.dart';
 import 'package:word_english/common/theme/custom_theme.dart';
 import 'package:word_english/screen/s_main.dart';
 
 class WordEnglishApp extends StatefulWidget {
+  static GlobalKey<NavigatorState> navigatorKey = GlobalKey();
+
   const WordEnglishApp({super.key});
 
   @override
   State<WordEnglishApp> createState() => _WordEnglishAppState();
 }
 
-class _WordEnglishAppState extends State<WordEnglishApp> {
+class _WordEnglishAppState extends State<WordEnglishApp> with Nav {
+
+  @override
+  GlobalKey<NavigatorState> get navigatorKey => WordEnglishApp.navigatorKey;
+
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
+      navigatorKey: navigatorKey,
       debugShowCheckedModeBanner: false,
       themeMode: ThemeMode.system,
       theme: lightTheme,
@@ -20,6 +28,8 @@ class _WordEnglishAppState extends State<WordEnglishApp> {
       home: const MainScreen(),
     );
   }
+
+
 }
 
 ThemeData lightTheme = ThemeData(
